Change representation from single tree to sets of trees
The main change in this commit is that `ClientTree` and `ServerTree`
now no longer just store a single tree and it's linearization, rather
it stores set (`[]`) of trees with this linearization.

I have not as of yet tested this feature in a situation where multiple
trees actually have the same linearization.

This commit also adds some better error reporting to the `FromJSON`
instance of `TTree`'s.