Issue #34: [WIP] Implement precompute subcommand sans binary instance for `Grammar`
The challenge here is that there is no `Binary` instance for `PGF` (a
projection of `Grammar`).  The implementors of `pgf` have decided to
copy the source code of `binary` and build against that rather than
submitting PR's for the functionality they needed against upstream.

There is a flag in the `pgf` library to not use their custom `Binary`
class, but I haven't had any success with building it with this option.