Make error reporting more robust
This commit heavily reduces the amount of calls to `error`.  It also
includes some other goodies like more heavy usage of `Text` rather
than `String`.  I've also make a monad transformer for the database
monad and also made the protocol monad a newtype.  Hence the blob of
newtype derived instances where it is declared.

This commit also fixes a bug introduced in
497dd4bf58322d4e39046c7974a3f0664a276ce2 where the time it took to
solve an exercise was serialized incorrectly.