Persist appstate across whole uptime
This is a rather large change to accomodate the fact that we want to
read the grammar files from the file system.  This is done by lazily
reading the grammars from the XDG_DATA_DIR as they are
encountered (actually most of them are generated at app-start because
we build a context objects from the grammars when we start the
database).

In the same go that I implemented this I noticed that some resources
where relinquished and rebuilt on each subsequent request.  This has
now been fixed and all data is maintained across requests.